cli/operator: extract assemble() + assembled + Close() (10b)
Move run()'s wiring (keys through operator.New) into assemble(ctx, cfg, logger, res, networkConfig, beacon, el) (*assembled, error) in assemble.go: it constructs the node components and returns them in the assembled struct. On failure it closes whatever it opened (db/p2p) via error-only defers keyed on the named return; on success the assembled owns those closers, torn down by Close() (p2p then db). The beacon/EL clients are injected as interfaces (beaconClient / executionclient.Provider) — the testability seam for the commit-11 smoke test.

run() now builds the clients (EL reordered up, as it needs the registry address) + networkConfig, then calls assemble + defers Close; the servers+start region stays inline (reading the assembled's fields) and is extracted into assembled.start() in 10c.
